Add files via upload
This tool is used for finding keyword in one file. It can search all directory which you have the rights. It is designed by thinking that when we need find some defining in one file ..for example we want file "EOF" keyword ,but I do not konw which file it is in. so，you can use this tool" bash search.sh /usr/include EOF"....hopeful nicing your work.